+++ apps/forms/components/hooks/use-dialog-state.ts
@@ -2,9 +2,9 @@ import { useState } from "react";
 
 export function useDialogState<T = any>(
   name = "dialog",
-  config?: { refreshkey?: boolean }
+  config?: { refreshkey?: boolean; defaultOpen?: boolean }
 ) {
-  const [open, setOpen] = useState(false);
+  const [open, setOpen] = useState<boolean>(config?.defaultOpen ?? false);
   const [key, setKey] = useState(0);
   const [data, setData] = useState<T>();
   const openDialog = (data?: T | Event) => {
