+++ apps/forms/scaffolds/blocks-editor/blocks/html-block.tsx
@@ -16,10 +16,8 @@ import {
   useDeleteBlock,
 } from "./base-block";
 import { useEditorState } from "@/scaffolds/editor";
-import { Editor, useMonaco } from "@monaco-editor/react";
 import { Button } from "@/components/ui/button";
-import { useMonacoTheme } from "@/components/monaco";
-import { useTheme } from "next-themes";
+import { ThemedMonacoEditor, useMonacoTheme } from "@/components/monaco";
 
 export function HtmlBlock({ id, body_html }: EditorFlatFormBlock) {
   const [state, dispatch] = useEditorState();
@@ -36,10 +34,6 @@ export function HtmlBlock({ id, body_html }: EditorFlatFormBlock) {
   );
 
   const deleteBlock = useDeleteBlock();
-
-  const monaco = useMonaco();
-  const { resolvedTheme } = useTheme();
-  useMonacoTheme(monaco, resolvedTheme ?? "light");
 
   return (
     <FlatBlockBase
@@ -88,7 +82,7 @@ export function HtmlBlock({ id, body_html }: EditorFlatFormBlock) {
       </BlockHeader>
       <div>
         <div className="rounded overflow-hidden border aspect-auto">
-          <Editor
+          <ThemedMonacoEditor
             height={400}
             defaultLanguage="html"
             value={body_html ?? ""}
